+++ src/core/analyzer.ts
@@ -20,8 +20,6 @@ import {Analysis, Document, Warning} from '../model/model';
 import {PackageRelativeUrl, ResolvedUrl} from '../model/url';
 import {Parser} from '../parser/parser';
 import {Scanner} from '../scanning/scanner';
-import {FsUrlLoader} from '../url-loader/fs-url-loader';
-import {PackageUrlResolver} from '../url-loader/package-url-resolver';
 import {UrlLoader} from '../url-loader/url-loader';
 import {UrlResolver} from '../url-loader/url-resolver';
 
@@ -85,20 +83,6 @@ export class Analyzer {
       this._urlLoader = context.loader;
       this._analysisComplete = Promise.resolve(context);
     }
-  }
-
-  /**
-   * Returns an analyzer with configuration inferred for the given directory.
-   *
-   * Currently this just creates a simple analyzer with a Fs loader rooted
-   * at the given directory, but in the future it may take configuration from
-   * files including polymer.json or similar.
-   */
-  static createForDirectory(dirname: string): Analyzer {
-    return new Analyzer({
-      urlLoader: new FsUrlLoader(dirname),
-      urlResolver: new PackageUrlResolver({packageDir: dirname})
-    });
   }
 
   /**