+++ src/test/html/html-document_test.ts
@@ -18,22 +18,22 @@ import * as fs from 'fs';
 import * as parse5 from 'parse5';
 import * as path from 'path';
 
-import {Analyzer} from '../../core/analyzer';
 import {ParsedHtmlDocument} from '../../html/html-document';
 import {HtmlParser} from '../../html/html-parser';
 import {PackageUrlResolver} from '../../url-loader/package-url-resolver';
-import {CodeUnderliner, fixtureDir} from '../test-utils';
+import {createForDirectory, fixtureDir} from '../test-utils';
 
-suite('ParsedHtmlDocument', () => {
+suite('ParsedHtmlDocument', async () => {
   const parser: HtmlParser = new HtmlParser();
   const url = `./source-ranges/html-complicated.html`;
   const file = fs.readFileSync(path.join(fixtureDir, url), 'utf8');
-  const analyzer = Analyzer.createForDirectory(fixtureDir);
-  const document: ParsedHtmlDocument = parser.parse(
-      file,
-      analyzer.resolveUrl(url)!,
-      new PackageUrlResolver({packageDir: fixtureDir}));
-  const underliner = new CodeUnderliner(analyzer);
+  const {analyzer, underliner} = await createForDirectory(fixtureDir);
+
+  let document: ParsedHtmlDocument;
+  setup(() => {
+    document =
+        parser.parse(file, analyzer.resolveUrl(url)!, analyzer.urlResolver);
+  });
 
   suite('sourceRangeForNode()', () => {
     test('works for comments', async () => {